requirements-dev.txt
# harness-boot — development dependencies
#
# Install with:
#   python -m pip install --user -r requirements-dev.txt
# or (if PEP 668 blocks it on your system Python):
#   python -m pip install --break-system-packages -r requirements-dev.txt
#
# CI (.github/workflows/self-check.yml) installs this set so the local
# matrix matches CI. Keep the two in sync — a missing dep that only hides
# test failures locally is what surfaced the v0.8.4 bug.

pyyaml
jsonschema
pytest       # v0.8.10 — required by tests/unit/test_gate_runner.py::PytestCommandDetectionTests
coverage     # v0.8.10 — enables gate_3 coverage detection tests via pytest-cov path
tomli; python_version < "3.11"  # v0.11.4 — tomllib backport for Python 3.10 CI matrix
